pyproject.toml
[tool.poetry]
name = "web-project-group-9"
version = "0.1.0"
description = ""
authors = ["AndriKovalchuk <andrii.kovalchuk90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
sqlalchemy = "^2.0.29"
cloudinary = "^1.40.0"
requests = "^2.31.0"
django-redis = "^5.4.0"
pillow = "^10.3.0"
django-environ = "^0.11.2"
websockets = "^12.0"
channels = "^4.1.0"
daphne = "^4.1.2"
channels-redis = "^4.2.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
